Add diagnostics for performance optimization (#4)
Add diagnostics for performance optimization. In the web app I've added basic elapsed time logging. I also created a new command-line app that has parameters for defining the fractal bounds and can be run to generate a standard Go pprof file. I included a helper script to run this in profile mode and generate PDFs. One known limitation is that the command-line app only supports float64 parameters for now, which is sufficient for the performance tuning I need to do but doesn't demonstrate full application functionality.