Fill window while maintaining aspect ratio (#11)
Update fractal and canvas resize code so that the window is consistently filled with a fractal of the correct aspect ratio, both upon initial load as well as subsequent window resizes. If the window is shrunk or the aspect ratio is changed then make sure we continue to show at least as much of the fractal as before. As part of this work, I switched the fractal and canvas representation to be the center point and a range.