Render fractal in tiles instead of a single image (#13)
Render fractal in tiles instead of a single image, which enables better responsiveness and opportunity for optimization. Use divs for rows and cells within each row. Switched to using background-image for displaying the fractal tiles.

In this current implementation, cells are all the same size which may leave a border around the fractal if it is not precisely divisible by the cell size.