Iteratively load the tiles with increasing resolution (#14)
Update tile loading so that they start with a low-res version of the fractal and gradually get more precise. Added a "grain" parameter to the backend that allows user to specify precision of fractal rendering in terms of pixels. Then implemented frontend loading of increasingly detailed grains (8 -> 4 -> 2 -> 1) with increasing intervals.